I think that is a good way to reach people you really want to follow.
In previous version, it would follow if user's ratio
les than min_previous_ratio. I.E. Follow = 100, Following = 10000
its ratio will be 0.01. I think following that user is not a good idea.

So basically, we will be able to control the scale that we really want.
It works properly for me.

If you accept this change, there should be a field in filter.json
"max_potency_ratio":

{
    "skip_business": true,
    "skip_non_business": false,
    "min_followers": 30,
    "max_followers": 1000,
    "min_followings": 30,
    "max_followings": 1000,
    "min_potency_ratio": 0.3,
    "max_potency_ratio": 1,
    "follow_private_or_empty": true
}

this is my settings, by looking that, it wont follow a user which has
300 follower, 1000 followings -> ratio 0.3