random double-click improvement
just an easy improvement..
I've realized that the double click was on the middle screen, no matter what! it looks like that this method doesn't have the offset parameter because it is not related to the UI object!! so I decided to change the logic and made two single clicks with a random delay.. :)

Other info:
You can only pass x, y and time between the clicks in double_click().
If you don't you get an error: "double_click() missing 2 required positional arguments: 'x' and 'y'"

Recap:
With double_click(x,y) you can do a double click on the screen in a coordinate you pass as argument.
With click() it press on the UI object center, with offset I'm able to press in all the right area without problems.

That's the reason I think we should use two click(offset=..) instead of a double_click(x+offset, y+offset) which in original code was a double_click on the middle of the screen.