double click fix + debug
Find out that make 2 single click in a row sometimes doens't work properly:t he time of the execution can be very high!
So we went back to double_click method and improve that.
We also added some debug code in order to track possibile clicks outside the UI objects

Co-Authored-By: Arthur Silva <arthursilva.ee@gmail.com>