Scroll when like btn not fully visible (#68)
* Only access like button if it exists

Check if the like button exists first
before trying to get its bounds

* Avoid double clicking close to the borders

Some posts have Reels link or tagged users link
in the bottom left corner. Now the double click
randomly click in an area inside of the selected
element but with a 30% padding from all sides
to avoid clicking in undesired objects.

* Scroll down when like btn is not fully visible

Sometimes the like button is in the view but
obscured by the tab bar. This tries to
scroll down in that case as well.

Co-authored-by: Arthur Silva <arthursilva.ee@gmail.com>