Use username for log files
Initialiy it will create a temporary log file
with an uuid. As soon as the username is
fetched, it will rename start writting to the
log file with username. Any initial log will be
moved from the temp log file to the new log
file with username.

All crash files will have the latest log file.