Filter added - min_posts
But there is a problem. I didn't check if account is private or not. It took too much time.
This way it is showing private accounts posts 0. If you have an idea let me know.
I put the filter after related to followings, business and before the ones related to alphabet (to make it faster)