scrapping thing
new plugins that allows user to generate txt instead of interact with users.
You have only to add that key argument to run a screaping session.

TODO: extend to others plugin, for now I've inplemented only for actions-bloggers
improve the verbose: current status should't not display how many likes, follow and so on but only the number of scraping targets.